+++ src/piglet.rs
@@ -27,47 +27,15 @@ use tracing_subscriber::EnvFilter;
 use hw::config::HardwareConfig;
 use hw::Hardware;
 
-#[cfg(feature = "iroh")]
-use crate::hw::PIGLET_ALPN;
-
-#[cfg(feature = "iroh")]
-use iroh_net::{
-    key::SecretKey,
-    relay::{RelayMode, RelayUrl},
-    Endpoint, NodeId,
-};
-
 use derive_more::Display;
-#[cfg(feature = "iroh")]
-use futures_lite::StreamExt;
-#[cfg(feature = "iroh")]
-use log::error;
 use serde::{Deserialize, Serialize};
 use std::fmt::Formatter;
 use std::{fs::File, io::Write};
 
-mod hardware_helper;
 mod hw;
-const SERVICE_NAME: &str = "net.mackenzie-serres.pigg.piglet";
-
 #[cfg(feature = "iroh")]
-#[derive(Serialize, Deserialize)]
-struct IrohInfo {
-    pub nodeid: NodeId,
-    pub local_addrs: String,
-    pub relay_url: RelayUrl,
-    pub alpn: String,
-    #[serde(skip)]
-    pub endpoint: Option<Endpoint>,
-}
-
-impl Display for IrohInfo {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        writeln!(f, "nodeid: {}", self.nodeid)?;
-        writeln!(f, "relay URL: {}", self.relay_url)?;
-        Ok(())
-    }
-}
+mod iroh_helper;
+const SERVICE_NAME: &str = "net.mackenzie-serres.pigg.piglet";
 
 #[cfg(feature = "tcp")]
 #[derive(Serialize, Deserialize)]
@@ -87,7 +55,7 @@ impl Display for TcpInfo {
 #[derive(Serialize, Deserialize)]
 struct ListenerInfo {
     #[cfg(feature = "iroh")]
-    pub iroh_info: Option<IrohInfo>,
+    pub iroh_info: Option<iroh_helper::IrohInfo>,
     #[cfg(feature = "tcp")]
     pub tcp_info: Option<TcpInfo>,
 }
@@ -159,7 +127,7 @@ async fn run_service(info_path: &Path, matches: &ArgMatches) -> anyhow::Result<(
     };
 
     let listener_info = ListenerInfo {
-        iroh_info: get_iroh_listener_info().await.ok(),
+        iroh_info: iroh_helper::get_iroh_listener_info().await.ok(),
         tcp_info: get_tcp_listener_info().await.ok(),
     };
 
@@ -176,7 +144,7 @@ async fn run_service(info_path: &Path, matches: &ArgMatches) -> anyhow::Result<(
     #[cfg(feature = "iroh")]
     if let Some(iroh_info) = listener_info.iroh_info {
         if let Some(endpoint) = iroh_info.endpoint {
-            listen_iroh(&endpoint, &mut hw).await?;
+            iroh_helper::listen_iroh(&endpoint, &mut hw).await?;
         }
     }
 
@@ -301,114 +269,6 @@ async fn listen_tcp(_tcp_info: TcpInfo, _hardware: &mut impl Hardware) -> anyhow
     Ok(())
 }
 
-#[cfg(feature = "iroh")]
-async fn get_iroh_listener_info() -> anyhow::Result<IrohInfo> {
-    let secret_key = SecretKey::generate();
-
-    // Build a `Endpoint`, which uses PublicKeys as node identifiers, uses QUIC for directly
-    // connecting to other nodes, and uses the relay protocol and relay servers to holepunch direct
-    // connections between nodes when there are NATs or firewalls preventing direct connections.
-    // If no direct connection can be made, packets are relayed over the relay servers.
-    let endpoint = Endpoint::builder()
-        // The secret key is used to authenticate with other nodes.
-        // The PublicKey portion of this secret key is how we identify nodes, often referred
-        // to as the `node_id` in our codebase.
-        .secret_key(secret_key)
-        // set the ALPN protocols this endpoint will accept on incoming connections
-        .alpns(vec![PIGLET_ALPN.to_vec()])
-        // `RelayMode::Default` means that we will use the default relay servers to holepunch and relay.
-        // Use `RelayMode::Custom` to pass in a `RelayMap` with custom relay urls.
-        // Use `RelayMode::Disable` to disable holepunching and relaying over HTTPS
-        // If you want to experiment with relaying using your own relay server,
-        // you must pass in the same custom relay url to both the `listen` code AND the `connect` code
-        .relay_mode(RelayMode::Default)
-        // pass in `0` to bind the socket to a random available port
-        .bind(0)
-        .await?;
-
-    let nodeid = endpoint.node_id();
-    println!("nodeid: {nodeid}");
-
-    let local_addrs = endpoint
-        .direct_addresses()
-        .next()
-        .await
-        .context("no endpoints")?
-        .into_iter()
-        .map(|endpoint| endpoint.addr.to_string())
-        .collect::<Vec<_>>()
-        .join(" ");
-    println!("local Addresses: {local_addrs}");
-
-    let relay_url = endpoint
-        .home_relay()
-        .expect("should be connected to a relay server, try calling `endpoint.local_endpoints()` or `endpoint.connect()` first, to ensure the endpoint has actually attempted a connection before checking for the connected relay server");
-    println!("Relay URL: {relay_url}");
-
-    Ok(IrohInfo {
-        nodeid,
-        local_addrs,
-        relay_url,
-        alpn: String::from_utf8_lossy(PIGLET_ALPN).parse()?,
-        endpoint: Some(endpoint),
-    })
-}
-
-#[cfg(feature = "iroh")]
-/// Listen for an incoming iroh-net connection and apply any config changes received, and
-/// send to GUI over the connection any input level changes.
-/// This is adapted from the iroh-net example with help from the iroh community
-async fn listen_iroh(endpoint: &Endpoint, hardware: &mut impl Hardware) -> anyhow::Result<()> {
-    loop {
-        // accept incoming connections, returns a normal QUIC connection
-        if let Some(connecting) = endpoint.accept().await {
-            let connection = connecting.await?;
-            let node_id = iroh_net::endpoint::get_remote_node_id(&connection)?;
-            info!("New connection from nodeid: '{node_id}'",);
-
-            let mut gui_sender = connection.open_uni().await?;
-
-            trace!("Sending hardware description");
-            let desc = hardware.description()?;
-            let message = serde_json::to_string(&desc)?;
-            gui_sender.write_all(message.as_bytes()).await?;
-            gui_sender.finish().await?;
-
-            loop {
-                trace!("waiting for connection");
-                match connection.accept_uni().await {
-                    Ok(mut config_receiver) => {
-                        let connection_clone = connection.clone();
-                        trace!("Connected, waiting for message");
-                        let payload = config_receiver.read_to_end(4096).await?;
-
-                        if !payload.is_empty() {
-                            let content = String::from_utf8_lossy(&payload);
-                            if let Ok(config_message) = serde_json::from_str(&content) {
-                                if let Err(e) = hardware_helper::apply_config_change(
-                                    hardware,
-                                    config_message,
-                                    connection_clone,
-                                )
-                                .await
-                                {
-                                    error!("Error applying config to hw: {}", e);
-                                }
-                            } else {
-                                error!("Unknown message: {content}");
-                            };
-                        }
-                    }
-                    _ => {
-                        info!("Connection lost");
-                        break;
-                    }
-                }
-            }
-        }
-    }
-}
-
 #[cfg(any(feature = "iroh", feature = "tcp"))]
 /// Write info about the running piglet to the info file
 fn write_info_file(info_path: &Path, listener_info: &ListenerInfo) -> anyhow::Result<()> {
@@ -496,7 +356,7 @@ mod test {
     use std::path::PathBuf;
     use std::str::FromStr;
 
-    use crate::{IrohInfo, ListenerInfo};
+    use crate::ListenerInfo;
     use tempfile::tempdir;
 
     #[cfg(feature = "iroh")]
@@ -512,7 +372,7 @@ mod test {
             .expect("Could not create Relay URL");
 
         let info = ListenerInfo {
-            iroh_info: Some(IrohInfo {
+            iroh_info: Some(crate::iroh_helper::IrohInfo {
                 nodeid,
                 local_addrs: local_addrs.to_string(),
                 relay_url,
@@ -543,7 +403,7 @@ mod test {
         let relay_url = iroh_net::relay::RelayUrl::from_str("https://euw1-1.relay.iroh.network./ ")
             .expect("Could not create Relay URL");
         let info = ListenerInfo {
-            iroh_info: Some(IrohInfo {
+            iroh_info: Some(crate::iroh_helper::IrohInfo {
                 nodeid,
                 local_addrs: local_addrs.to_string(),
                 relay_url,
